+++ application/build.gradle
@@ -1,8 +1,11 @@
 dependencies {
     implementation project(path: ':menu')
     implementation project(path: ':order')
+    implementation project(path: ':order_table')
     implementation project(path: ':product')
-    implementation project(path: ':common')
+
+    implementation 'org.flywaydb:flyway-core'
+    runtimeOnly 'com.h2database:h2'
 }
 
 tasks.named('bootJar') {
