+++ build.gradle
@@ -22,11 +22,8 @@ subprojects {
     apply { plugin('java') }
 
     dependencies {
-        implementation 'org.springframework.boot:spring-boot-starter-actuator'
         implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
         implementation 'org.springframework.boot:spring-boot-starter-web'
-        implementation 'org.flywaydb:flyway-core'
-        runtimeOnly 'com.h2database:h2'
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
     }
 
@@ -37,7 +34,6 @@ subprojects {
     tasks.named('jar') {
         enabled = true
     }
-
 
     tasks.named('test') {
         useJUnitPlatform()
