+++ http/table-groups.http
@@ -1,6 +1,6 @@
 ###
 POST {{host}}/api/table-groups
-Content-Type: application/json
+Content-Type: kitchenpos.application/json
 
 {
   "orderTableIds": [1, 2]
