--- order/src/main/java/kitchenpos/application/OrderService.java
@@ -1,52 +1,0 @@
-package kitchenpos.application;
-
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.repository.OrderRepository;
-import kitchenpos.domain.repository.OrderTableRepository;
-import kitchenpos.dto.response.OrderResponse;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-public class OrderService {
-    private final OrderRepository orderRepository;
-    private final OrderTableRepository orderTableRepository;
-
-    public OrderService(OrderTableRepository orderTableRepository, OrderRepository orderRepository) {
-        this.orderRepository = orderRepository;
-        this.orderTableRepository = orderTableRepository;
-    }
-
-    @Transactional
-    public Long create(final Long orderTableId) {
-        OrderTable orderTable = orderTableRepository.getById(orderTableId);
-        Order order = new Order(orderTable, OrderStatus.COOKING, LocalDateTime.now());
-        return orderRepository.save(order).getId();
-    }
-
-    public List<OrderResponse> list() {
-        final List<Order> orders = orderRepository.findAll();
-        return orders.stream()
-                .map(OrderResponse::from)
-                .collect(Collectors.toList());
-    }
-
-    @Transactional
-    public OrderResponse changeOrderStatus(final Long orderId, final OrderStatus orderStatus) {
-        final Order savedOrder = orderRepository.findById(orderId)
-                .orElseThrow(IllegalArgumentException::new);
-
-        savedOrder.validateStatusIsEqualToCompletion();
-
-        savedOrder.updateOrderStatus(orderStatus);
-        orderRepository.save(savedOrder);
-        return OrderResponse.from(savedOrder);
-    }
-
-}
