--- order/src/main/java/kitchenpos/domain/OrderStatus.java
@@ -1,5 +1,0 @@
-package kitchenpos.domain;
-
-public enum OrderStatus {
-    COOKING, MEAL, COMPLETION
-}
