--- order/src/main/java/kitchenpos/domain/OrderTableGroup.java
@@ -1,43 +1,0 @@
-package kitchenpos.domain;
-
-import javax.persistence.Embedded;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import java.time.LocalDateTime;
-
-import static javax.persistence.GenerationType.IDENTITY;
-
-@Entity
-public class OrderTableGroup {
-
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-
-    @Embedded
-    private OrderTables orderTables;
-
-    private LocalDateTime createdDate;
-
-    protected OrderTableGroup() {
-
-    }
-
-    public OrderTableGroup(OrderTables orderTables, LocalDateTime createdDate) {
-        this.orderTables = orderTables;
-        this.createdDate = createdDate;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
-    public OrderTables getOrderTables() {
-        return orderTables;
-    }
-}
