--- order/src/main/java/kitchenpos/domain/repository/OrderRepository.java
@@ -1,16 +1,0 @@
-package kitchenpos.domain.repository;
-
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import java.util.List;
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-
-    @Query("SELECT CASE WHEN COUNT(o) > 0 THEN TRUE ELSE FALSE END FROM Order o WHERE o.orderTable.id IN :orderTableIds AND o.orderStatus IN :orderStatuses")
-    boolean existsByOrderTableIdInAndOrderStatusIn(@Param("orderTableIds") List<Long> orderTableIds, @Param("orderStatuses") List<OrderStatus> orderStatuses);
-
-}
