--- order/src/main/java/kitchenpos/dto/request/OrderStatusChangeRequest.java
@@ -1,20 +1,0 @@
-package kitchenpos.dto.request;
-
-import kitchenpos.domain.OrderStatus;
-
-public class OrderStatusChangeRequest {
-
-    private OrderStatus orderStatus;
-
-    public OrderStatusChangeRequest() {
-    }
-
-    public OrderStatusChangeRequest(OrderStatus orderStatus) {
-        this.orderStatus = orderStatus;
-    }
-
-    public OrderStatus getOrderStatus() {
-        return orderStatus;
-    }
-
-}
