--- order/src/main/java/kitchenpos/dto/request/OrderTableCreateRequest.java
@@ -1,24 +1,0 @@
-package kitchenpos.dto.request;
-
-public class OrderTableCreateRequest {
-
-    private int numberOfGuests;
-    private boolean empty;
-
-    public OrderTableCreateRequest() {
-    }
-
-    public OrderTableCreateRequest(int numberOfGuests, boolean empty) {
-        this.numberOfGuests = numberOfGuests;
-        this.empty = empty;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-
-    public boolean isEmpty() {
-        return empty;
-    }
-
-}
