--- order/src/main/java/kitchenpos/dto/request/TableGroupCreateRequest.java
@@ -1,20 +1,0 @@
-package kitchenpos.dto.request;
-
-import java.util.List;
-
-public class TableGroupCreateRequest {
-
-    private List<Long> orderTableIds;
-
-    public TableGroupCreateRequest() {
-    }
-
-    public TableGroupCreateRequest(List<Long> orderTableIds) {
-        this.orderTableIds = orderTableIds;
-    }
-
-    public List<Long> getOrderTableIds() {
-        return orderTableIds;
-    }
-
-}
