--- order/src/main/java/kitchenpos/dto/response/OrderTableResponse.java
@@ -1,37 +1,0 @@
-package kitchenpos.dto.response;
-
-import kitchenpos.domain.OrderTable;
-
-public class OrderTableResponse {
-
-    private final Long id;
-    private final int numberOfGuests;
-    private final boolean empty;
-
-    public OrderTableResponse(Long id, int numberOfGuests, boolean empty) {
-        this.id = id;
-        this.numberOfGuests = numberOfGuests;
-        this.empty = empty;
-    }
-
-    public static OrderTableResponse from(final OrderTable orderTable) {
-        return new OrderTableResponse(
-                orderTable.getId(),
-                orderTable.getNumberOfGuests(),
-                orderTable.isEmpty()
-        );
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-
-    public boolean isEmpty() {
-        return empty;
-    }
-
-}
