+++ order/src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,0 +1,47 @@
+package kitchenpos.order.application;
+
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.repository.OrderRepository;
+import kitchenpos.order.dto.response.OrderResponse;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+public class OrderService {
+    private final OrderRepository orderRepository;
+
+    public OrderService(OrderRepository orderRepository) {
+        this.orderRepository = orderRepository;
+    }
+
+    @Transactional
+    public Long create() {
+        Order order = new Order(OrderStatus.COOKING, LocalDateTime.now());
+        return orderRepository.save(order).getId();
+    }
+
+    public List<OrderResponse> list() {
+        final List<Order> orders = orderRepository.findAll();
+        return orders.stream()
+                .map(OrderResponse::from)
+                .collect(Collectors.toList());
+    }
+
+    @Transactional
+    public OrderResponse changeOrderStatus(final Long orderId, final OrderStatus orderStatus) {
+        final Order savedOrder = orderRepository.findById(orderId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        savedOrder.validateStatusIsEqualToCompletion();
+
+        savedOrder.updateOrderStatus(orderStatus);
+        orderRepository.save(savedOrder);
+        return OrderResponse.from(savedOrder);
+    }
+
+}
