+++ order/src/main/java/kitchenpos/order/domain/Order.java
@@ -1,0 +1,62 @@
+package kitchenpos.order.domain;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+import java.util.Objects;
+
+import static javax.persistence.EnumType.STRING;
+import static javax.persistence.GenerationType.IDENTITY;
+
+@Entity
+@Table(name = "orders")
+public class Order {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Long id;
+    private Long orderTableId;
+    @Enumerated(STRING)
+    private OrderStatus orderStatus;
+    private LocalDateTime orderedTime;
+
+    protected Order() {
+    }
+
+    public Order(OrderStatus orderStatus, LocalDateTime orderedTime) {
+        this.orderStatus = orderStatus;
+        this.orderedTime = orderedTime;
+    }
+
+    public Order(Long orderTableId, OrderStatus orderStatus, LocalDateTime orderedTime) {
+        this.orderTableId = orderTableId;
+        this.orderStatus = orderStatus;
+        this.orderedTime = orderedTime;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public Long getOrderTableId() {
+        return orderTableId;
+    }
+
+    public OrderStatus getOrderStatus() {
+        return orderStatus;
+    }
+
+    public LocalDateTime getOrderedTime() {
+        return orderedTime;
+    }
+
+    public void updateOrderStatus(final OrderStatus orderStatus) {
+        this.orderStatus = orderStatus;
+    }
+
+    public void validateStatusIsEqualToCompletion() {
+        if (Objects.equals(this.orderStatus, OrderStatus.COMPLETION)) {
+            throw new IllegalArgumentException("이미 완료된 주문의 상태는 변경할 수 없습니다!");
+        }
+    }
+
+}
