+++ order_table/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+    implementation project(path: ':order')
+}
