+++ order_table/src/main/java/kitchenpos/order_table/dto/request/OrderTableCreateRequest.java
@@ -1,0 +1,24 @@
+package kitchenpos.order_table.dto.request;
+
+public class OrderTableCreateRequest {
+
+    private int numberOfGuests;
+    private boolean empty;
+
+    public OrderTableCreateRequest() {
+    }
+
+    public OrderTableCreateRequest(int numberOfGuests, boolean empty) {
+        this.numberOfGuests = numberOfGuests;
+        this.empty = empty;
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+
+}
