+++ order_table/src/main/java/kitchenpos/order_table/dto/request/TableGroupCreateRequest.java
@@ -1,0 +1,20 @@
+package kitchenpos.order_table.dto.request;
+
+import java.util.List;
+
+public class TableGroupCreateRequest {
+
+    private List<Long> orderTableIds;
+
+    public TableGroupCreateRequest() {
+    }
+
+    public TableGroupCreateRequest(List<Long> orderTableIds) {
+        this.orderTableIds = orderTableIds;
+    }
+
+    public List<Long> getOrderTableIds() {
+        return orderTableIds;
+    }
+
+}
