+++ order_table/src/main/java/kitchenpos/order_table/dto/response/OrderTableResponse.java
@@ -1,0 +1,36 @@
+package kitchenpos.order_table.dto.response;
+
+import kitchenpos.order_table.domain.OrderTable;
+public class OrderTableResponse {
+
+    private final Long id;
+    private final int numberOfGuests;
+    private final boolean empty;
+
+    public OrderTableResponse(Long id, int numberOfGuests, boolean empty) {
+        this.id = id;
+        this.numberOfGuests = numberOfGuests;
+        this.empty = empty;
+    }
+
+    public static OrderTableResponse from(final OrderTable orderTable) {
+        return new OrderTableResponse(
+                orderTable.getId(),
+                orderTable.getNumberOfGuests(),
+                orderTable.isEmpty()
+        );
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+
+}
